+++ config/settings/production.py
@@ -10,6 +10,9 @@ ALLOWED_HOSTS = env.list("DJANGO_ALLOWED_HOSTS", default=["https://tele.band"])
 
 # DATABASES
 # ------------------------------------------------------------------------------
+DATABASES = {
+    "default": env.db("DATABASE_URL", default="postgres:///teleband"),
+}
 DATABASES["default"] = env.db("DATABASE_URL")  # noqa F405
 DATABASES["default"]["ATOMIC_REQUESTS"] = True  # noqa F405
 DATABASES["default"]["CONN_MAX_AGE"] = env.int("CONN_MAX_AGE", default=60)  # noqa F405
