+++ teleband/submissions/models.py
@@ -31,7 +31,7 @@ class Submission(models.Model):
         null=True,
         related_name="own_submission",
     )
-    assignment = models.ForeignKey(Assignment, on_delete=models.PROTECT)
+    assignment = models.ForeignKey(Assignment, on_delete=models.PROTECT, related_name="submissions")
     submitted = models.DateTimeField(auto_now_add=True)
     content = models.TextField(blank=True)
